Create README.md
In a major school there are students who have access to a study room, theoretically they go there to study but when more than a certain number of students come together they have a party. The school principal does not agree that the room should be used for parties. The rules of the simulation are as follows:
- There can be any number of students in the study room at the same time.
- The students only go to the room once a day (simulation).
- The director can only enter the room when there is no one there or when the number of students is greater than or equal to the MAX number allowed. When the number is lower, he does not enter the room so as not to disturb the students who are working there and stays there waiting to control what is going on inside. That is to say
o When there is no one there, it passes by.
o When there are students but they don't reach the MAX, it stays at the door waiting for them.
o When there is the MAX or more students enter the room and start the party.
- When the director is inside the study room, no more students can enter.
- The director does not leave the room until the last student has left.
- There is only one director and he makes three rounds a day to control the study room. of days there are holidays and he doesn't know it.